+++ src/features/Conversation/Plugins/Render/utils/pluginState.ts
@@ -1,4 +1,4 @@
-import { PluginChannel } from '@lobehub/chat-plugin-sdk/client';
+import { PluginChannel } from '@aipmorg/chat-plugin-sdk/client';
 import { useEffect } from 'react';
 
 export const useOnPluginStateUpdate = (callback: (key: string, value: any) => void) => {