+++ src/services/__tests__/tool.test.ts
@@ -55,7 +55,7 @@ describe('ToolService', () => {
       const manifestUrl = 'http://fake-url.com/manifest.json';
 
       const fakeManifest = {
-        $schema: '../node_modules/@lobehub/chat-plugin-sdk/schema.json',
+        $schema: '../node_modules/@aipmorg/chat-plugin-sdk/schema.json',
         api: [
           {
             url: 'https://realtime-weather.chat-plugin.lobehub.com/api/v1',
@@ -186,7 +186,7 @@ describe('ToolService', () => {
         const openapiUrl = 'http://fake-url.com/openapiUrl.json';
 
         const fakeManifest = {
-          $schema: '../node_modules/@lobehub/chat-plugin-sdk/schema.json',
+          $schema: '../node_modules/@aipmorg/chat-plugin-sdk/schema.json',
           api: [],
           openapi: openapiUrl,
           author: 'LobeHub',
@@ -223,7 +223,7 @@ describe('ToolService', () => {
         const openapiUrl = 'http://fake-url.com/openapiUrl.json';
         const manifestUrl = 'http://fake-url.com/manifest.json';
         const fakeManifest = {
-          $schema: '../node_modules/@lobehub/chat-plugin-sdk/schema.json',
+          $schema: '../node_modules/@aipmorg/chat-plugin-sdk/schema.json',
           api: [],
           openapi: openapiUrl,
           author: 'LobeHub',