+++ src/store/chat/slices/plugin/action.ts
@@ -1,5 +1,5 @@
 /* eslint-disable sort-keys-fix/sort-keys-fix, typescript-sort-keys/interface */
-import { PluginErrorType } from '@lobehub/chat-plugin-sdk';
+import { PluginErrorType } from '@aipmorg/chat-plugin-sdk';
 import { t } from 'i18next';
 import { Md5 } from 'ts-md5';
 import { StateCreator } from 'zustand/vanilla';