+++ src/store/tool/slices/store/initialState.ts
@@ -1,4 +1,4 @@
-import { LobeChatPluginMeta } from '@lobehub/chat-plugin-sdk';
+import { LobeChatPluginMeta } from '@aipmorg/chat-plugin-sdk';
 
 export type PluginInstallLoadingMap = Record<string, boolean | undefined>;